requirements.txt
flask
flask-cors
numpy
pillow
torch
torchvision
scikit-learn
requests
google-generativeai